refactor chunk
